Migrate delete-history commit test off the gci e2e suite
The scenario now fits the harness's nested-transaction commit
strategy: delete-history involves no instance migration, so
there is no persistence question. Moves it to
refactoringInstVar.committing.integration.test.ts, which runs in CI
across the release matrix, and updates the gci e2e suite's header
comment to describe only what's left there (migrate-instances and
the two accessor-atomicity scenarios) instead of the whole set this
plan already relocated.
